Create Blinkin Lights
I don't know what this is ;-;

This is the same base code last year. I will find the link to the video where I got it from, but I need to find it first...

I am trying to work on the sensor part of it, and it will take a while for me to work on it because I never got to touch that part last year. If you guys can help me code this, that will be great! :)

Signed-off-by: younge143 <younge143@carlsbadusd.net>